+++ apps/agent/pkg/cluster/interface.go
@@ -2,6 +2,7 @@ package cluster
 
 type Cluster interface {
 	Join(addrs []string) (clusterSize int, err error)
+	SyncMembership() error
 	Shutdown() error
 	FindNodes(key string, n int) ([]Node, error)
 	FindNode(key string) (Node, error)